refactor(team): extract team-settings + agent-status from index
Continues the team/index split (WS6). Pulls the two cohesive helper groups the
REST routes lean on hardest out of the 1,914-line index into lib/ modules:

- lib/team-settings.ts — display-settings + OrgTeam list I/O over the plugin
  settings store, plus reportsTo normalize/degrade and mergeDisplayDefaults.
  These are the most route-referenced helpers (readTeams x13, readDisplaySettings
  x8, writeDisplaySettings x5, mergeDisplayDefaults x5). Pure store-backed —
  zero plugin-context coupling.
- lib/agent-status.ts — online/working/offline status resolution (heartbeat +
  audit-activity signals) and org structure. The staleness-threshold context is
  now a `setStaleSettingsContext(ctx)` setter the plugin calls from activate(),
  replacing the bare module-level `staleSettingsCtx = ctx` assignment.

index.ts: 1,914 -> 1,685 lines. Pure relocation — populateTeamRoutes and
activate() now import these helpers instead of calling module-scope functions.

This also de-risks the remaining ~970-line populateTeamRoutes extraction: the
helper groups it depended on are now importable libs, leaving only the live
search-indexing wiring (pluginCtx/indexAgentStatic/batchIndexAgents) in index.

Verified: typecheck + lint clean; team suite 183/0; full suite 5123/0; binary
build; isolated-home boot smoke (Team activates; GET /api/plugins/team/ and
/teams both 200 — exercising resolveAgentStatus + display/team settings).
